+++ openeogeotrellis/job_registry.py
@@ -645,8 +645,9 @@ class InMemoryJobRegistry(JobRegistryInterface):
     def set_status(
         self,
         job_id: str,
-        status: str,
         *,
+        user_id: Optional[str] = None,
+        status: str,
         updated: Optional[str] = None,
         started: Optional[str] = None,
         finished: Optional[str] = None,
@@ -675,14 +676,20 @@ class InMemoryJobRegistry(JobRegistryInterface):
     def set_dependency_usage(self, job_id: str, *, user_id: Optional[str] = None, dependency_usage: Decimal) -> None:
         self._update(job_id, dependency_usage=str(dependency_usage))
 
-    def set_proxy_user(self, job_id: str, proxy_user: str) -> None:
+    def set_proxy_user(self, job_id: str, *, user_id: Optional[str] = None, proxy_user: str) -> None:
         self._update(job_id=job_id, proxy_user=proxy_user)
 
-    def set_application_id(self, job_id: str, application_id: str) -> None:
+    def set_application_id(self, job_id: str, *, user_id: Optional[str] = None, application_id: str) -> None:
         self._update(job_id=job_id, application_id=application_id)
 
     def set_results_metadata(
-        self, job_id: str, costs: Optional[float], usage: dict, results_metadata: Dict[str, Any]
+        self,
+        job_id: str,
+        *,
+        user_id: Optional[str] = None,
+        costs: Optional[float],
+        usage: dict,
+        results_metadata: Dict[str, Any],
     ) -> None:
         self._update(job_id=job_id, costs=costs, usage=usage, results_metadata=results_metadata)
 
@@ -870,14 +877,24 @@ class DoubleJobRegistry:  # TODO: extend JobRegistryInterface?
         elif zk_job_info is None and ejr_job_info is None:
             raise JobNotFoundException(job_id=job_id)
 
-    def set_status(self, job_id: str, user_id: str, status: str,
-                   started: Optional[str] = None, finished: Optional[str] = None,
-                   ) -> None:
+    def set_status(
+        self,
+        job_id: str,
+        *,
+        user_id: Optional[str] = None,
+        status: str,
+        updated: Optional[str] = None,
+        started: Optional[str] = None,
+        finished: Optional[str] = None,
+    ) -> None:
         if self.zk_job_registry:
+            assert user_id, "user_id is required in ZkJobRegistry"
             self.zk_job_registry.set_status(job_id=job_id, user_id=user_id, status=status, started=started,
                                             finished=finished)
         if self.elastic_job_registry:
-            self.elastic_job_registry.set_status(job_id=job_id, status=status, started=started, finished=finished)
+            self.elastic_job_registry.set_status(
+                job_id=job_id, user_id=user_id, status=status, updated=updated, started=started, finished=finished
+            )
 
     def delete_job(self, job_id: str, *, user_id: Optional[str] = None) -> None:
         if self.zk_job_registry:
@@ -922,24 +939,20 @@ class DoubleJobRegistry:  # TODO: extend JobRegistryInterface?
                 job_id=job_id, user_id=user_id, dependency_usage=dependency_usage
             )
 
-    def set_proxy_user(self, job_id: str, user_id: str, proxy_user: str) -> None:
+    def set_proxy_user(self, job_id: str, *, user_id: Optional[str] = None, proxy_user: str) -> None:
         # TODO: add dedicated method
         if self.zk_job_registry:
+            assert user_id, "user_id is required in ZkJobRegistry"
             self.zk_job_registry.patch(job_id=job_id, user_id=user_id, proxy_user=proxy_user)
         if self.elastic_job_registry:
-            self.elastic_job_registry.set_proxy_user(
-                job_id=job_id, proxy_user=proxy_user
-            )
+            self.elastic_job_registry.set_proxy_user(job_id=job_id, user_id=user_id, proxy_user=proxy_user)
 
-    def set_application_id(
-        self, job_id: str, user_id: str, application_id: str
-    ) -> None:
+    def set_application_id(self, job_id: str, *, user_id: Optional[str] = None, application_id: str) -> None:
         if self.zk_job_registry:
+            assert user_id, "user_id is required in ZkJobRegistry"
             self.zk_job_registry.set_application_id(job_id=job_id, user_id=user_id, application_id=application_id)
         if self.elastic_job_registry:
-            self.elastic_job_registry.set_application_id(
-                job_id=job_id, application_id=application_id
-            )
+            self.elastic_job_registry.set_application_id(job_id=job_id, user_id=user_id, application_id=application_id)
 
     def mark_ongoing(self, job_id: str, user_id: str) -> None:
         if self.zk_job_registry:
@@ -1026,12 +1039,20 @@ class DoubleJobRegistry:  # TODO: extend JobRegistryInterface?
             )
 
     def set_results_metadata(
-        self, job_id, user_id, costs: Optional[float], usage: dict, results_metadata: Dict[str, Any]
+        self,
+        job_id: str,
+        *,
+        user_id: Optional[str] = None,
+        costs: Optional[float],
+        usage: dict,
+        results_metadata: Dict[str, Any],
     ) -> None:
         if self.zk_job_registry:
+            assert user_id, "user_id is required in ZkJobRegistry"
             self.zk_job_registry.patch(job_id=job_id, user_id=user_id,
                                        **dict(results_metadata, costs=costs, usage=usage))
 
         if self.elastic_job_registry:
-            self.elastic_job_registry.set_results_metadata(job_id=job_id, costs=costs, usage=usage,
-                                                           results_metadata=results_metadata)
+            self.elastic_job_registry.set_results_metadata(
+                job_id=job_id, user_id=user_id, costs=costs, usage=usage, results_metadata=results_metadata
+            )